update
I have now fixed the INT berry situation with the great help of Whaea Munireh the GOAT of coding, ily Whaea !!!!!

both the amount of berries(if over 3 or under 1 would not work.) that has been fixed! (it was 'or' when it should have been 'and')

And the INT loop was not working!
I would explain what we did but it is too much. basically my int was both in the wrong spot and my while was also weird, fixed and working now !